pyproject.toml
[project]
name = "ecom-llm-cs"
version = "0.1.0"
description = "电商客服智能问答项目 - 大模型指令微调与偏好对齐"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.0.0",
    "transformers>=4.57.3",
    "trl>=0.8.5",
    "peft>=0.5.0",
    "torch==2.9.1",
    "accelerate>=0.22.0",
    "pandas>=2.0",
    "numpy>=1.25",
    "tqdm>=4.66",
    "openai>=1.0.0",          # DeepSeek API客户端
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0"
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Torch GPU/CPU选项
[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "gpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[project.optional-dependencies]
cpu = ["torch==2.9.1"]
gpu = ["torch==2.9.1"]

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "gpu" }
    ]
]